Implement migration of legacy Git-Mastery metadata to new folder structure
