+++ app/configs/gitmastery_config.py
@@ -1,12 +1,11 @@
 import json
-import os
-import shutil
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Self, Type, Optional, Union
 
 from app.configs.utils import read_config
 
+OLD_GITMASTERY_CONFIG_NAME = ".gitmastery.json"
 METADATA_FOLDER_NAME = ".gitmastery"
 GITMASTERY_CONFIG_NAME = "config.json"
 GITMASTERY_LOG_NAME = "gitmastery.log"
@@ -114,30 +113,3 @@ GIT_MASTERY_EXERCISES_SOURCE = GitMasteryConfig.ExercisesSource.from_raw(
         "branch": "main",
     }
 )
-
-
-def migrate_gitmastery_metadata(root: Path) -> None:
-    from app.commands.progress.constants import PROGRESS_LOCAL_FOLDER_NAME
-    from app.utils.cli import rmtree
-
-    gitmastery_dir = root / METADATA_FOLDER_NAME
-    gitmastery_dir.mkdir(exist_ok=True)
-    shutil.copy2(root / ".gitmastery.json", gitmastery_dir / GITMASTERY_CONFIG_NAME)
-    legacy_log = root / ".gitmastery.log"
-    new_log = gitmastery_dir / GITMASTERY_LOG_NAME
-    if legacy_log.exists():
-        shutil.copy2(legacy_log, new_log)
-    else:
-        new_log.touch()
-    if (root / PROGRESS_LOCAL_FOLDER_NAME).is_dir() and not (
-        gitmastery_dir / PROGRESS_LOCAL_FOLDER_NAME
-    ).exists():
-        shutil.copytree(
-            root / PROGRESS_LOCAL_FOLDER_NAME,
-            gitmastery_dir / PROGRESS_LOCAL_FOLDER_NAME,
-        )
-    os.remove(root / ".gitmastery.json")
-    if legacy_log.exists():
-        os.remove(legacy_log)
-    if (root / PROGRESS_LOCAL_FOLDER_NAME).is_dir():
-        rmtree(root / PROGRESS_LOCAL_FOLDER_NAME)
