+++ app/hooks/in_gitmastery_root.py
@@ -6,9 +6,10 @@ import click
 from app.configs.gitmastery_config import (
     GITMASTERY_CONFIG_NAME,
     METADATA_FOLDER_NAME,
+    OLD_GITMASTERY_CONFIG_NAME,
     GitMasteryConfig,
-    migrate_gitmastery_metadata,
 )
+from app.configs.migration import migrate_gitmastery_metadata
 from app.configs.utils import find_root
 from app.hooks.utils import generate_cds_string
 from app.utils.click import CliContextKey, error, warn
@@ -34,21 +35,25 @@ def in_gitmastery_root(
         ) -> Any:
             root = find_root(GITMASTERY_CONFIG_NAME, folder=METADATA_FOLDER_NAME)
             if root is None:
-                old_root = find_root(".gitmastery.json")
+                old_root = find_root(OLD_GITMASTERY_CONFIG_NAME)
+
+                # User is not in a Git-Mastery root folder
                 if old_root is None:
                     error(
                         f"You are not in a Git-Mastery root folder. Navigate to an appropriate folder or use "
                         f"{click.style('gitmastery setup', bold=True, italic=True)}"
                     )
+
+                # User has old metadata structure, attempt to migrate to new structure
                 try:
                     migrate_gitmastery_metadata(old_root[0])
                     warn("Migrated your Git-Mastery metadata to .gitmastery/ folder.")
                     root = find_root(
                         GITMASTERY_CONFIG_NAME, folder=METADATA_FOLDER_NAME
                     )
-                except Exception:
-                    error(MIGRATION_FAILURE_MESSAGE)
-                if root is None:
+                    if root is None:
+                        error(MIGRATION_FAILURE_MESSAGE)
+                except (FileNotFoundError, PermissionError, OSError):
                     error(MIGRATION_FAILURE_MESSAGE)
 
             path, cds = root
